refactor: SettlementWithItems 클래스 batch 패키지로 이동 - #122
